network: introduce hardware timestamping
Some network adapters support hardware timestamping of packets sent
and/or received. Since this is implemented in hardware, timestamping can
only be configured globally for all applications. In addition there are
restrictions as to _when_ this configuration can be applied.
Consequently it is convenient to let systemd-networkd configure hardware
timestamping instead of leaving it up to some other service.

This commit adds a new `[HardwareTimestamping]` section to .network
files with two settings: `ReceiveMode` and `TransmitMode`. These
settings use the same names that `ethtool -T <iface>` emits.